- FieldType tests

+++ backend/src/field_types/field_type.service.ts
@@ -1,9 +1,8 @@
-import { Injectable, HttpException, HttpStatus } from "@nestjs/common";
+import { Injectable, HttpException, HttpStatus, NotFoundException, ConflictException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { FieldType } from "./field_type.entity";
 import { Repository } from "typeorm";
 import { FieldTypeDto } from "./field_type.dto";
-import { ValidationService } from "src/validations/validation.service";
 
 @Injectable()
 export class FieldTypeService {
@@ -12,16 +11,29 @@ export class FieldTypeService {
         private readonly type_repo: Repository<FieldType>
     ) { }
 
-    async findAll(): Promise<FieldType[]> {
-        return await this.type_repo.find()
+    async findAll(): Promise<Object> {
+        return { items: await this.type_repo.find() }
     }
 
     async findById(id: string): Promise<FieldType> {
-        return await this.type_repo.findOne({ id: id })
+        try {
+            let type = await this.type_repo.findOneOrFail({ id: id })
+            return type
+        } catch (e) {
+            throw new NotFoundException()
+        }
     }
 
     async deleteType(id: string) {
+        let toDelete = await this.type_repo.findOne(id)
+
+        if (!toDelete) {
+            throw new NotFoundException()
+        }
+
         await this.type_repo.delete({ id: id })
+
+        return {}
     }
 
     async addType(data: FieldTypeDto): Promise<FieldType> {
@@ -29,6 +41,10 @@ export class FieldTypeService {
 
         try {
             type = await this.type_repo.save(type)
+
+            // Making a find so we are returning the full object (including relations)
+            return await this.type_repo.findOne(type.id)
+
         } catch (e) {
             if (e.code && e.code == 23505) {
                 throw new HttpException(e.detail, HttpStatus.CONFLICT)
@@ -38,21 +54,34 @@ export class FieldTypeService {
             }
         }
 
-        return type
     }
 
     async updateType(id: string, data: FieldTypeDto): Promise<FieldType> {
+        if (data.name) {
+            // Prevent duplicated names
+            let duplicated = await this.type_repo.createQueryBuilder('t')
+                .select("t.id")
+                .where("t.name = :name AND t.id != :id ", { name: data.name, id: id })
+                .getRawOne()
+
+            if (duplicated) {
+                throw new ConflictException(`There is an already existing field type with the name: ${data.name}`)
+            }
+        }
+
         data['id'] = id;
         let type = await this.type_repo.preload(data)
-        // overriding the incoming validations (since validations are a relation the preload do a merge instead of a replacement)
-        type.validations = data.validations;
 
         if (!type) {
             throw new HttpException('Unable to find the required type', HttpStatus.NOT_FOUND)
         }
 
+        // overriding the incoming validations (since validations are a relation the preload do a merge instead of a replacement)
+        type.validations = data.validations;
         type = await this.type_repo.save(type)
-        return type
+
+        // Making a find so we are returning the full object (including relations)
+        return await this.type_repo.findOne(id)
     }
 
     async getIDsFromNames(names: string[]): Promise<Object[]> {
